src/main/java/kz/home/ys/algo/medium/oddEvenLinkedList/OddEvenLinkedListLeetCode.java
package kz.home.ys.algo.medium.oddEvenLinkedList;

class OddEvenLinkedListLeetCode {

    // time - O(N)
    // space - O(1)
    public ListNode oddEvenList(ListNode head) {
        if (head == null) return null;

        ListNode odd = head;
        ListNode even = head.next;
        ListNode evenHead = even;
        while (even != null && even.next != null) {
            odd.next = even.next;
            odd = odd.next;
            even.next = odd.next;
            even = even.next;
        }
        odd.next = evenHead;

        return head;
    }
}











